Add GameCamera Class

+++ CatsAdventure/CatsAdventure/Bullet.h
@@ -4,7 +4,7 @@
 const Vector2f bulletBitmapScale = { 40.f, 40.f };
 const Vector2f bulletColliderScale = { 20.f, 20.f };
 const float bulletSpeed = 800.f;
-const float bulletDamage = 2.f;
+const int bulletDamage = 2;
 
 class Bullet :
     public Projectile
