+++ CatsAdventure/CatsAdventure/Web.h
@@ -4,7 +4,7 @@
 const Vector2f webBitmapScale = { 30.f, 30.f };
 const Vector2f webColliderScale = { 15.f, 15.f };
 const float webSpeed = 400.f;
-const float webDamage = 20.f;
+const int webDamage = 20;
 
 class Web :
 	public Projectile
